refactor: extract cut detection into useRewardCutHistory hook
Move data fetching and malicious behavior detection out of the component
into a dedicated hook. Detection is now directional: only flags
increases (bad for delegators) of 50+ percentage points, not decreases
(which benefit delegators). An orch dropping from 100% to 0% reward cut
is positive and won't trigger a warning; raising from 0% to 50%+ will.
